Move dashboard home to its own folder
