+++ components/dashboard/dashboard-home/dashboard-banner/banner-image.jsx
