+++ components/dashboard/dashboard-home/dashboard-banner/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Icon, Tag } from 'antd';
 import Link from 'next/link';
-import RoundButton from '../../../../common/buttons/round-button';
-import FollowStationButton from '../../../../common/buttons/follow-station-button';
+import RoundButton from '../../../common/buttons/round-button';
+import FollowStationButton from '../../../common/buttons/follow-station-button';
 import BannerContainer from './banner-container';
 import BannerImage from './banner-image';
 import BannerCover from './banner-cover';
@@ -16,7 +16,11 @@ export default ({ station }) => (
       <Tag color="#333">FEATURED</Tag>
       <h1>{station.name}</h1>
       {station.description ? <h2>{station.description}</h2> : null}
-      <Link href={`/stations/${station.slug}`}>
+      <Link
+        prefetch
+        as={`/stations/${station.slug}`}
+        href={`/stations?slug=${station.slug}`}
+      >
         <RoundButton type="primary">
           Listen <Icon type="arrow-right" />
         </RoundButton>
